Updating sortType of SHOW_SETS_QUERY to RECENT from STANDARD (depreciated)
The STANDARD sort-type seemed to be causing issues with returning repeated sets from subsequent function calls with iterating page numbers. This behavior was noticed in the tournament_show_sets function, but this led to the discovery that STANDARD is depreciated in the startgg API and a replacement should be used instead.